+++ source/E2E.Tests/Util/ObjectBuilders/ItemObjectBuilder.cs
@@ -3,10 +3,8 @@
 namespace E2E.Tests.Util.ObjectBuilders;
 internal class ItemObjectBuilder : IObjectBuilder
 {
-    private static int itemCounter;
-
     public object Build()
     {
-        return new ItemObject($"ItemObject_Tests_{++itemCounter}");
+        return new ItemObject();//new ItemObject(Guid.NewGuid().ToString());
     }
 }
