+++ source/GameInterface/AutoSync/Templates/Handlers/SubscribeSetReferenceTemplate.txt
@@ -1,17 +1,17 @@
 ﻿{{~ if Interface ~}}
-Subscribe<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetLocalMessage>((instanceId, data) =>
+Subscribe<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetLocalMessage>((instanceId, data) => 
 {
     int index = AutoSync.Constants.TypeMap[data.Value.GetType()];
     if (!this.objectManager.TryGetId(data.Value.GetType(), data.Value, out string valueId) && data.Value != null) return;
-
-    network.SendAll(new {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetNetworkMessage(instanceId, valueId, index));
+                
+    network.SendAll(new {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetNetworkMessage(instanceId, valueId, index));
 });
 
-SubscribeNetwork<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetNetworkMessage>((instance, data) =>
+SubscribeNetwork<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetNetworkMessage>((instance, data) => 
 {
     var type = AutoSync.Constants.TypeMapReverse[data.ValueTypeIndex];
     if (!objectManager.TryGetObjectWithLogging(type, data.ValueId, out object value) && data.ValueId != null) return;
-{{~ if ReflectionAssignment ~}}
+{{~ if ReadOnly ~}}
 	AccessTools.Field(typeof({{MemberDeclaringType}}), "{{MemberName}}").SetValue(instance, value);
 {{~ else ~}}
 	using (new AllowedThread())
@@ -21,14 +21,14 @@ SubscribeNetwork<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberIdentifie
 {{~ end ~}}
 });
 {{~ else ~}}
-SubscribeGenericReference<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetLocalMessage, {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetNetworkMessage>();
+SubscribeGenericReference<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetLocalMessage, {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetNetworkMessage>();
 
-SubscribeNetworkReference<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetNetworkMessage >((instance, value, _) => {
+SubscribeNetworkReference<{{MemberType}}, {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetNetworkMessage >((instance, value, _) => { 
 {{~ if Debug ~}}
 	Logger.Debug("[Client] Received network set message for {type}", "{{MemberName}}");
 {{~ end ~}}
 
-    {{~ if ReflectionAssignment ~}}
+    {{~ if ReadOnly ~}}
         AccessTools.Field(typeof({{MemberDeclaringType}}), "{{MemberName}}").SetValue(instance, value);
     {{~ else ~}}
         using (new AllowedThread())
