+++ source/GameInterface/AutoSync/Templates/Messages/LocalSetMessageTemplate.txt
@@ -5,9 +5,9 @@ using {{ library }};
 
 namespace AutoSync;
 
-public record {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetLocalMessage : GenericEvent<{{ MemberDeclaringType }}, {{ MemberType }}>
+public record {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetLocalMessage : GenericEvent<{{ MemberDeclaringType }}, {{ MemberType }}>
 {
-    public {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetLocalMessage({{ MemberDeclaringType }} instance, {{ MemberType }} value) : base(instance, value)
+    public {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetLocalMessage({{ MemberDeclaringType }} instance, {{ MemberType }} value) : base(instance, value)
     {
     }
 }
