+++ source/GameInterface/AutoSync/Templates/Messages/NetworkSetReferenceMessageTemplate.txt
@@ -10,7 +10,7 @@ using {{ library }};
 namespace AutoSync
 {
     [ProtoContract(SkipConstructor = true)]
-    public record {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetNetworkMessage : GenericNetworkReferenceEvent<{{ MemberDeclaringType}}, {{ MemberType }}>
+    public record {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetNetworkMessage : GenericNetworkReferenceEvent<{{ MemberDeclaringType}}, {{ MemberType }}>
     {
         [ProtoMember(1)]
         public override string InstanceId { get; set; }
@@ -18,7 +18,7 @@ namespace AutoSync
         [ProtoMember(2)]
         public override string ValueId { get; set; }
 
-        public {{ MemberDeclaringTypeName}}_{{ MemberIdentifier }}_SetNetworkMessage(string instanceId, string valueId) : base(instanceId, valueId)
+        public {{ MemberDeclaringTypeName}}_{{ MemberName }}_SetNetworkMessage(string instanceId, string valueId) : base(instanceId, valueId)
         {
         }
     }
