+++ source/GameInterface/Services/ItemRosters/Handlers/UpdateItemRosterHandler.cs
@@ -1,7 +1,6 @@
 ﻿using Common;
 using Common.Logging;
 using Common.Messaging;
-using GameInterface.Services.ItemObjects;
 using GameInterface.Services.ItemRosters.Messages;
 using GameInterface.Services.ItemRosters.Patches;
 using GameInterface.Services.ObjectManager;
@@ -20,13 +19,11 @@ internal class UpdateItemRosterHandler : IHandler
     private static readonly ILogger Logger = LogManager.GetLogger<UpdateItemRosterHandler>();
     private readonly IMessageBroker messageBroker;
     private readonly IObjectManager objectManager;
-    private readonly ItemObjectRegistry itemObjectRegistry;
 
-    public UpdateItemRosterHandler(IMessageBroker messageBroker, IObjectManager objectManager, ItemObjectRegistry itemObjectRegistry)
+    public UpdateItemRosterHandler(IMessageBroker messageBroker, IObjectManager objectManager)
     {
         this.messageBroker = messageBroker;
         this.objectManager = objectManager;
-        this.itemObjectRegistry = itemObjectRegistry;
 
         messageBroker.Subscribe<UpdateItemRoster>(Handle);
     }
@@ -44,7 +41,7 @@ internal class UpdateItemRosterHandler : IHandler
         {
             try
             {
-                if (TryGetItem(msg.ItemId, out ItemObject item) == false)
+                if (objectManager.TryGetObject(msg.ItemId, out ItemObject item) == false)
                 {
                     Logger.Error("Unable to find item with id: {itemId}", msg.ItemId);
                     return;
@@ -69,14 +66,6 @@ internal class UpdateItemRosterHandler : IHandler
                 Logger.Error(e, "Failed to apply {Message}", nameof(UpdateItemRoster));
             }
         });
-    }
-
-    private bool TryGetItem(string itemId, out ItemObject item)
-    {
-        if (objectManager.TryGetObject(itemId, out item))
-            return true;
-
-        return itemObjectRegistry.TryGetRegisteredItem(itemId, out item);
     }
 
     public void Dispose()
