+++ source/GameInterface/Services/MapEvents/Handlers/ConversationRequestHandler.cs
@@ -4,7 +4,6 @@ using Common.Messaging;
 using Common.Network;
 using Common.Network.Messages;
 using Common.Util;
-using GameInterface.Services.MapEvents;
 using GameInterface.Services.MapEvents.Messages;
 using GameInterface.Services.MapEvents.Messages.Conversation;
 using GameInterface.Services.MapEvents.PlayerPartyInteractions;
@@ -143,9 +142,7 @@ internal class ConversationRequestHandler : IHandler
         // Mark and hold on the game thread, then reply, so the party is frozen and protected before the client
         // (re)opens the encounter. The filter above runs on the network thread; game state can change while the
         // approval is queued, so the hold step re-validates everything it depends on.
-        GameThread.RunSafe(
-            () => HoldAndApprove(requestingPeer, request, aiPartyId, playerPartyId),
-            context: nameof(Handle_NetworkRequestConversation));
+        GameThread.Run(() => HoldAndApprove(requestingPeer, request, aiPartyId, playerPartyId));
     }
 
     /// <summary>
@@ -362,7 +359,7 @@ internal class ConversationRequestHandler : IHandler
             return;
         }
 
-        GameThread.RunSafe(() =>
+        GameThread.Run(() =>
         {
             try
             {
@@ -402,7 +399,7 @@ internal class ConversationRequestHandler : IHandler
                 Logger.Error(e, "Failed to restart approved conversation encounter; releasing the server-side party hold");
                 SendConversationEndedToServer();
             }
-        }, context: nameof(Handle_NetworkAllowConversation));
+        });
     }
 
     /// <summary>[Client] This player's encounter finished; tell the server to release the held party.</summary>
@@ -441,9 +438,7 @@ internal class ConversationRequestHandler : IHandler
     {
         if (ModInformation.IsServer) return;
 
-        GameThread.RunSafe(
-            ConversationPartyHold.ShowInteractionBlockedMessage,
-            context: nameof(Handle_NetworkConversationDenied));
+        GameThread.Run(ConversationPartyHold.ShowInteractionBlockedMessage);
     }
 
     /// <summary>[Server] The defender's client reports it is showing the "hold on" popup; record its peer so a
@@ -483,8 +478,7 @@ internal class ConversationRequestHandler : IHandler
     /// <summary>[Server] Releases the given player's engagement on the game thread.</summary>
     private void ReleaseEngagementOnMainThread(NetPeer peer)
     {
-        GameThread.RunSafe(() =>
-            ConversationPartyHold.EndEngagement(conversationPartyTracker, peer),
-            context: nameof(ReleaseEngagementOnMainThread));
+        GameThread.Run(() =>
+            ConversationPartyHold.EndEngagement(conversationPartyTracker, peer));
     }
 }
